requirements.txt
# requirements.txt

tensorflow
pandas
librosa
matplotlib
transformers
torch
torchaudio
tqdm
audiomentations==0.33.0
scikit-learn
soundfile
speechbrain
pyannote.metrics